+++ lib/smart_core/engine.rb
@@ -8,4 +8,5 @@ module SmartCore::Engine
   require_relative 'engine/rescue_ext'
   require_relative 'engine/atom'
   require_relative 'engine/frozener'
+  require_relative 'engine/cache'
 end